+++ APL-Matrix-Calculator/CppLibrary/CppLibrary.h
@@ -10,8 +10,8 @@
 /*
 * Add two matrices
 */
-extern "C" CPPLIBRAR_API void cppAdd(int* A, int* B, int* C, int rows, int columns);
+extern "C" CPPLIBRAR_API void cppAdd(float* A, float* B, float* C, int rows, int columns);
    
-extern "C" CPPLIBRAR_API void cppSub(int* A, int* B, int* C, int rows, int columns);
+extern "C" CPPLIBRAR_API void cppSub(float* A, float* B, float* C, int rows, int columns);
 
-extern "C" CPPLIBRAR_API void cppMul(int* A, int* B, int* C, int rows, int columns, int colsA, int colsB);
+extern "C" CPPLIBRAR_API void cppMul(float* A, float* B, float* C, int rows, int columns, int colsA, int colsB);
